move the ViewRegistry up one package, and clean up the visibility of other classes
