+++ sdk/src/jmh/java/io/opentelemetry/sdk/metrics/TestSdk.java
@@ -21,7 +21,6 @@ import io.opentelemetry.metrics.DefaultMeter;
 import io.opentelemetry.metrics.Meter;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.internal.MillisClock;
-import io.opentelemetry.sdk.metrics.view.ViewRegistry;
 import io.opentelemetry.sdk.resources.Resource;
 
 public enum TestSdk {
